+++ jest.config.js
@@ -38,10 +38,10 @@ module.exports = {
   // Configure coverage thresholds
   coverageThreshold: {
     global: {
-      branches: 70,
-      functions: 80,
-      lines: 80,
-      statements: 80
+      branches: 60,
+      functions: 60,
+      lines: 60,
+      statements: 60
     }
   },
   // Output verbose coverage info